misc/HtmlTestSuite/results/test-1302.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>origin of the script that invoked the method, javascript:</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><iframe src="javascript:''"/>
<div id="log"/>
<script>
async_test(function() {
  window[0].postMessage('', '*', []);
  window[0].onmessage = this.step_func(function(e) {
    assert_equals(e.origin, location.protocol + '//' + location.host);
    assert_array_equals(e.ports, []);
    this.done();
  });
});
</script>


</body></html>
